[SPARK-49868][K8S] Fix race condition in executor failure tracking
